feat: Support workflow_run pull requests
Due to security issues running untrusted checkouts from forks, access to things
like secrets and tokens are blocked. It is possible to allow this, but it is
not wise, and the recommended process is to use a workflow run call. This will
add support for workflow_run based pull requests, allowing PRs to be commented
upon, regardless of whether they are a fork or from a local branch.

fixes #259